fix: int64 index tensors + refine KV capture guard for decode
1. capture.py: change _pos_tensor/_count_tensor/_capacity_tensor
   from int32 to int64 — PyTorch index_copy_() requires long tensor
   for the index parameter.

2. vllm.py: refine _capture_kv guard:
   - T=1 (decode): always run write_graph — it IS graph-safe and must
     be captured INTO the CUDA graph for replay
   - T>1 (prefill): only run when not capturing — quantization creates
     new tensors which is forbidden during stream capture

+++ turboquant/capture.py
@@ -165,10 +165,10 @@ class RingBuffer:
 
         Call once during model initialisation, before any CUDA Graph capture.
         """
-        self._pos_tensor = torch.zeros(1, device=self.device, dtype=torch.int32)
-        self._count_tensor = torch.zeros(1, device=self.device, dtype=torch.int32)
+        self._pos_tensor = torch.zeros(1, device=self.device, dtype=torch.int64)
+        self._count_tensor = torch.zeros(1, device=self.device, dtype=torch.int64)
         self._capacity_tensor = torch.tensor(
-            [self.capacity], device=self.device, dtype=torch.int32,
+            [self.capacity], device=self.device, dtype=torch.int64,
         )
         self._arange_buf = torch.arange(
             self.capacity, device=self.device, dtype=torch.int32,
